// frontend/app/shared/api/models/IUserCreate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
// @ts-nocheck
export type IUserCreate = {
    email: string;
    password: string;
    is_active?: (boolean | null);
    is_superuser?: (boolean | null);
    is_verified?: (boolean | null);
    first_name?: (string | null);
    last_name?: (string | null);
    middle_name?: (string | null);
    phone?: (string | null);
    tg_id?: (string | null);
};